django-api/requirements.txt
Django
djangorestframework
django-filter
drf-spectacular
drf-spectacular-sidecar
drf-standardized-errors
whitenoise
dj-database-url
djangorestframework-simplejwt
psycopg2-binary
# django-redis
django-debug-toolbar
django-cors-headers
pymongo
# pika